fix: decouple `vane-portals` from `vane-regions`
