+++ vane-regions/src/main/java/org/oddlama/vane/regions/menu/RoleMenu.java
@@ -57,7 +57,7 @@ public class RoleMenu extends ModuleComponent<Regions> {
 	public TranslatedItemStack<?> item_setting_info_container;
 	public TranslatedItemStack<?> item_setting_info_portal;
 
-	public RoleMenu(Context<Regions> context) {
+	public RoleMenu(Context<Regions> context, boolean vane_portals_available) {
 		super(context.namespace("role"));
 		final var ctx = get_context();
 		item_rename = new TranslatedItemStack<>(ctx, "rename", Material.NAME_TAG, 1, "Used to rename the role.");
@@ -158,14 +158,17 @@ public class RoleMenu extends ModuleComponent<Regions> {
 				1,
 				"Used to represent the info for the container setting."
 			);
-		item_setting_info_portal =
-			new TranslatedItemStack<>(
-				ctx,
-				"setting_info_portal",
-				Material.ENDER_PEARL,
-				1,
-				"Used to represent the info for the portal setting."
-			);
+
+		if (vane_portals_available) {
+			item_setting_info_portal =
+					new TranslatedItemStack<>(
+							ctx,
+							"setting_info_portal",
+							Material.ENDER_PEARL,
+							1,
+							"Used to represent the info for the portal setting."
+					);
+		}
 	}
 
 	public Menu create(final RegionGroup group, final Role role, final Player player) {
@@ -192,7 +195,10 @@ public class RoleMenu extends ModuleComponent<Regions> {
 		add_menu_item_setting(role_menu, role, 2, item_setting_info_build, RoleSetting.BUILD);
 		add_menu_item_setting(role_menu, role, 4, item_setting_info_use, RoleSetting.USE);
 		add_menu_item_setting(role_menu, role, 5, item_setting_info_container, RoleSetting.CONTAINER);
-		add_menu_item_setting(role_menu, role, 8, item_setting_info_portal, RoleSetting.PORTAL);
+
+		if (item_setting_info_portal != null) {
+			add_menu_item_setting(role_menu, role, 8, item_setting_info_portal, RoleSetting.PORTAL);
+		}
 
 		role_menu.on_natural_close(player2 -> get_module().menus.region_group_menu.create(group, player2).open(player2)
 		);
